--- a/db/migrate/20151204171252_update_branch_name_for_pr_branches.rb
+++ b/db/migrate/20151204171252_update_branch_name_for_pr_branches.rb
@@ -1,0 +1,17 @@
+class UpdateBranchNameForPrBranches < ActiveRecord::Migration
+  class Branch < ActiveRecord::Base; end # Don't use the real model
+
+  def up
+    Branch.where(:pull_request => true).each do |branch|
+      pr_number = branch.name.split("/").last
+      branch.update_attributes(:name => "pulls/#{pr_number}/head")
+    end
+  end
+
+  def down
+    Branch.where(:pull_request => true).each do |branch|
+      pr_number = branch.name.split("/")[1]
+      branch.update_attributes(:name => "pr/#{pr_number}")
+    end
+  end
+end