--- a/spec/support/miq_tools_services_helper.rb
+++ b/spec/support/miq_tools_services_helper.rb
@@ -1,5 +1,7 @@
 def stub_git_service
   double("MiniGit service").tap do |git|
+    allow(git).to receive(:ensure_prs_refs)
+    allow(git).to receive(:fetch).with("--all")
     allow(MiqToolsServices::MiniGit).to receive(:call).and_yield(git)
   end
 end