Requires changes in https://github.com/ManageIQ/miq_tools_services/pull/19

- Don't create extra branches
- Use the new fetched PRs 
- Migrate PR Branches to use the new name