@jrafanie 2015-12-09
Looks good.  The initial commit monitor looks to be slower, but the average run seems to be 5 times faster.  Below is the top slowest PR/Commit Monitor jobs after letting them run from a clean repo.

**Before**:

```
2015-12-09T03:19:06.882Z 111875 TID-5zatg PullRequestMonitor JID-f2067042d78a35f40853f804 INFO: done: 242.318 sec
2015-12-09T03:18:54.974Z 111875 TID-tr3q0 CommitMonitor JID-e301a09064b115e3d27e0a6e INFO: done: 230.431 sec
2015-12-09T03:13:39.951Z 111875 TID-tr3q0 PullRequestMonitor JID-03e22b2d2eb5a67dc7959d04 INFO: done: 204.707 sec
2015-12-09T03:28:03.345Z 111875 TID-65soc PullRequestMonitor JID-f8b3cde7177209e3e88e9058 INFO: done: 192.746 sec
2015-12-09T03:27:56.170Z 111875 TID-tr2io CommitMonitor JID-b141f79e1abef6f3028ea258 INFO: done: 185.885 sec
2015-12-09T03:38:08.478Z 111875 TID-65ix8 CommitMonitor JID-5d76f3d9be7b0f40cf1439f1 INFO: done: 175.325 sec
2015-12-09T03:57:55.337Z 111875 TID-65jms PullRequestMonitor JID-6adb73243c5200f553cabb0d INFO: done: 174.224 sec
2015-12-09T03:47:51.564Z 111875 TID-tr50o PullRequestMonitor JID-04c9ac4ac5810342d2a9d229 INFO: done: 167.956 sec
2015-12-09T03:57:48.598Z 111875 TID-668hs CommitMonitor JID-d1092e5b2022e1c16af4dedc INFO: done: 167.488 sec
2015-12-09T04:07:48.150Z 111875 TID-65jms PullRequestMonitor JID-ecf567e780e2e412e2c5fbeb INFO: done: 164.218 sec
2015-12-09T03:47:44.822Z 111875 TID-65res CommitMonitor JID-339169d10721c4ac750dad48 INFO: done: 161.219 sec
2015-12-09T04:03:08.137Z 111875 TID-tr4c8 PullRequestMonitor JID-671e386467aa69a050cc993b INFO: done: 160.852 sec
2015-12-09T04:07:41.140Z 111875 TID-tr2pc CommitMonitor JID-ef2e62a1e354f7d50c24234d INFO: done: 157.239 sec
2015-12-09T03:52:41.117Z 111875 TID-65soc CommitMonitor JID-63bdd9ed3233b13d4f37edb9 INFO: done: 154.972 sec
2015-12-09T03:42:39.151Z 111875 TID-tr3d8 PullRequestMonitor JID-3260ce993fc988e3e114af19 INFO: done: 150.41 sec
2015-12-09T03:32:36.720Z 111875 TID-tr3q0 CommitMonitor JID-61b98f52a84303b16e277301 INFO: done: 149.886 sec
2015-12-09T04:02:56.590Z 111875 TID-tr41o CommitMonitor JID-da3c8b28deaef6658e448076 INFO: done: 149.313 sec
2015-12-09T03:42:26.258Z 111875 TID-65soc CommitMonitor JID-fec44dcf392794fcc0891ef2 INFO: done: 137.525 sec
2015-12-09T03:35:20.719Z 111875 TID-tr2io PullRequestMonitor JID-ba8b66258fa0176519097595 INFO: done: 7.56 sec
2015-12-09T03:30:13.999Z 111875 TID-tr41o PullRequestMonitor JID-fdacd38a999d431b1ccd8c3f INFO: done: 7.16 sec
2015-12-09T03:50:12.843Z 111875 TID-tr33s PullRequestMonitor JID-b20087d00f73becb03dd051d INFO: done: 6.696 sec
```

**After**:
```
2015-12-09T02:04:08.986Z 94368 TID-6srao CommitMonitor JID-c2d9bf5307e939ff95695526 INFO: done: 436.127 sec
2015-12-09T02:04:05.866Z 94368 TID-daoi0 PullRequestMonitor JID-e0258d1f0ca3ccad952f5408 INFO: done: 433.002 sec
2015-12-09T02:34:14.611Z 94368 TID-1tijps PullRequestMonitor JID-b06994f17b10d793d1235b2a INFO: done: 250.28 sec
2015-12-09T02:36:18.877Z 94368 TID-dao8k CommitMonitor JID-9fde65f6c9569d8f6b6a49e7 INFO: done: 62.112 sec
2015-12-09T02:30:55.712Z 94368 TID-daoi0 CommitMonitor JID-baddc73007336aacce67a9cb INFO: done: 51.547 sec
2015-12-09T02:06:08.648Z 94368 TID-6rl8s CommitMonitor JID-b9deb7baee30e6c0c7ebac1d INFO: done: 50.458 sec
2015-12-09T02:36:00.980Z 94368 TID-6srao PullRequestMonitor JID-a9cfe221069c4f2f36cef94b INFO: done: 44.487 sec
2015-12-09T02:35:51.581Z 94368 TID-danq8 CommitMonitor JID-ee9b38a0c29ece40e1a3c5b2 INFO: done: 35.325 sec
2015-12-09T02:55:40.007Z 94368 TID-6s204 CommitMonitor JID-9c054d24b1d79ef0785067e5 INFO: done: 34.047 sec
2015-12-09T03:00:34.478Z 94368 TID-1sxh2o PullRequestMonitor JID-401ce0b55632367f50e78de3 INFO: done: 27.729 sec
2015-12-09T02:40:52.366Z 94368 TID-6spkg CommitMonitor JID-88cfa7a1b08ac5afd7ba5c4a INFO: done: 26.619 sec
2015-12-09T02:50:46.735Z 94368 TID-celcg CommitMonitor JID-fd8eb28c9b2c9f7aa065e34d INFO: done: 26.571 sec
2015-12-09T02:35:38.991Z 94368 TID-dapos PullRequestMonitor JID-596232df2ef2be423de84242 INFO: done: 22.958 sec
2015-12-09T02:45:29.512Z 94368 TID-6rkaw PullRequestMonitor JID-c0efcf9c0ca79657cfb14308 INFO: done: 21.934 sec
2015-12-09T02:55:26.351Z 94368 TID-6sd8k PullRequestMonitor JID-93f7c6e63dad392cc07dbb40 INFO: done: 20.388 sec
2015-12-09T02:50:34.537Z 94368 TID-danq8 PullRequestMonitor JID-1e7a2ccca26a6c0970489281 INFO: done: 14.369 sec
2015-12-09T02:40:39.980Z 94368 TID-6sd8k PullRequestMonitor JID-a592ae5c2b2028bc0b8ede00 INFO: done: 14.229 sec
2015-12-09T02:35:29.130Z 94368 TID-cej9g CommitMonitor JID-4200c4af6a38b6a0b03af44d INFO: done: 13.327 sec
2015-12-09T02:45:20.452Z 94368 TID-6rl8s CommitMonitor JID-d644dcc38a9d47fbc4335929 INFO: done: 12.879 sec
2015-12-09T03:00:19.554Z 94368 TID-danq8 CommitMonitor JID-d4a652f88f3f45469b01246a INFO: done: 12.814 sec
```